# swagger.yaml
openapi: "3.0.0"

info:
  title: learn express and documentation
  description: Here is a segment where i am writing docs
  version: 1.1.0
  contact:
    email: vaibhavroykumar@gmail.com
    url: "https://vaibhavkumarroy.netlify.app/"

##servers:
##  - url: "https://localhost:4000/api/v1" 
##    description: for local host - secure
##  - url: "http://localhost:4000/api/v1" 
##    description: for local host - regular

servers:
  - url: "{protocol}://localhost:4000/api/{version}"
    description: for localhost - secure
    variables: 
      version:
        enum:
          - v1
          - v2 
        default: v1
      protocol:
        enum:
          - http
          - https 
        default: https

components:
  securitySchemes:
   cookieAuth:
    type: apiKey
    in: cookie
    name: token
   BearerAuth:
      type: http
      scheme: bearer


paths:
  /vaibhav:
    get:
     tags:
     - String
     summary: Returns a greet message from Vaibhav
     responses: 
      200:
        description: All good request
        content: 
          application/json:
            schema:
              type: string
              example: "mystring"
      400:
        description: bad request
      
      500:
        description: Internal Server error

  /vaibhavobject:
     get:
      tags:
      - Object
      summary: Returns a unique course
      responses: 
       200:
         description: All good request
         content: 
           application/json:
             schema:
               type: object
               properties:
                id: 
                 type: string
                name: 
                 type: string
                price: 
                 type: number
       400:
         description: bad request
      
       500:
         description: Internal Server error
  
  /courses:
     get:
      tags:
      - Array
      summary: Returns all courses
      responses: 
       200:
         description: All good request
         content: 
           application/json:
             schema:
               type: array
               items:
                type: object
                properties:
                  id: 
                   type: string
                  name: 
                   type: string
                  price: 
                   type: number
       400:
         description: bad request
      
       500:
         description: Internal Server error
    
  /mycourses/{courseId}:
     get:
      tags:
      - String
      summary: Returns selected course Id based on the URL
      parameters: 
       - name: courseId
         in: path
         required: true
         default: 22
         schema:
          type: string
      responses: 
       200:
         description: All good request
         content: 
           application/json:
             schema:
               type: object
               properties:
                 id: 
                  type: string
                 name: 
                  type: string
                 price: 
                  type: number
       400:
         description: bad request
      
       500:
         description: Internal Server error


  /addCourse:
     post:
      tags:
      - String
      summary: Add a new course into the existing courses
      consumes:
       - application/json
      produces:
       - application/json
      requestBody:
        required: true
        content: 
          application/json:
            schema:
              type: object
              properties:
                 id: 
                  type: string
                 name: 
                  type: string
                 price: 
                  type: number
      responses: 
       200:
         description: All good request
         content: 
           application/json:
             schema:
               type: boolean
       400:
         description: bad request
      
       500:
         description: Internal Server error


  /coursequery:
     get:
      tags:
      - String
      summary: trying to learn about query
      parameters:
       - name: location
         in: query
         required: true
         schema:
           type: string
           enum: [delhi , noida , gurgaon]
       - name: device
         in: query
         required: true
         schema:
           type: string
           enum: [web , mobile]
      responses: 
       200:
         description: All good request
         content: 
           application/json:
             schema:
               type: object
               properties:
                 location:
                   type: string
                 device:
                   type: string     
       400:
         description: bad request
      
       500:
         description: Internal Server error

  /courseUpload:
     post:
      tags:
      - String
      summary: trying to learn how to upload file through post
      parameters: 
       - in: header
         name: auth
      requestBody:
        content:
         multipart/form-data:
           schema:
            type: object
            properties:
              file:
                type: string
                format: binary
      responses: 
       200:
         description: All good request
       400:
         description: bad request
       500:
         description: Internal Server error